+++ shop_files/lang/german/extra/admin/kk_altcha_catcha.php
@@ -1,2 +1,11 @@
 <?php
+/* ------------------------------------------------------------
+  Module "ALTCHA - Captcha" made by Karl
+
+  modified eCommerce Shopsoftware
+  http://www.modified-shop.org
+
+  Released under the GNU General Public License
+-------------------------------------------------------------- */
+
 define('CFG_TXT_KK_ALTCHA', 'Altcha Captcha');