+++ shop_files/lang/german/extra/kk_altcha_catcha.php
@@ -8,4 +8,4 @@
   Released under the GNU General Public License
 -------------------------------------------------------------- */
 
-define('CFG_TXT_KK_ALTCHA', 'Altcha Captcha');
+define('KK_ALTCHA_CAPTCHA_JS_MISSING', 'Hinweis: F&uuml;r die Sicherheitspr&uuml;fung wird JavaScript ben&ouml;tigt. Bitte aktivieren Sie JavaScript in Ihrem Browser, um das Formular abzusenden.');